// src/community/ui/data/college/models/CollegeColor.ts
export enum CollegeColor {
  Default = 'navy',
  AI = 'purple',
  DT = 'violet',
  Global = 'green',
  Leadership = 'blue',
  Management = 'teal',
  SV = 'orange',
  Happiness = 'yellow',
  InnovationDesign = 'red',
  SemicondDesign = 'marine',
  EnergySolution = 'bluegreen',
  BMDesign = 'pink-sub',
}